+++ Sources/StreamChat/Database/DatabaseSession.swift
@@ -159,6 +159,8 @@ protocol ChannelDatabaseSession {
     
     /// Fetches `ChannelDTO` with the given `cid` from the database.
     func channel(cid: ChannelId) -> ChannelDTO?
+    
+    func deleteChannels(query: ChannelListQuery) throws
 }
 
 protocol ChannelReadDatabaseSession {